requirements.txt
absl-py==1.0.0
astunparse==1.6.3
black==21.12b0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
contourpy==1.0.6
cycler==0.11.0
flatbuffers==2.0
fonttools==4.38.0
gast==0.5.3
google-auth==2.4.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
imageio==2.14.1
isort==5.10.1
keras==2.8.0rc1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.6.2
mypy==0.931
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.22.1
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.1
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
PyWavelets==1.2.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.1
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tifffile==2021.11.2
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
